requirements.txt
discord
python-dotenv
google-cloud-firestore